chore: refactoring rfc.ts into multiple files. New RfcCommon type to unify Rfc and TypeSense data structures
