+++ client/components/RFCCardCompact.vue
@@ -12,7 +12,7 @@
     container-class="flex"
   >
     <template #headingTitle>
-      <component :is="formatTitle(`rfc${props.rfc.number}`)" />
+      <component :is="formatTitleAsVNode(`rfc${props.rfc.number}`)" />
     </template>
     <p
       class="ml-4 pl-4 border-l-1 border-gray-300 text-base text-blue-900 dark:text-white flex items-center"
@@ -23,14 +23,14 @@
 </template>
 
 <script setup lang="ts">
-import type { Rfc } from '../generated/red-client'
 import { infoRfcPathBuilder } from '../utilities/url'
-import { formatTitle } from '~/utilities/rfc'
+import { formatTitleAsVNode } from '~/utilities/rfc'
+import type { RfcCommon } from '~/utilities/rfc'
 import { useResponsiveModeStore } from '~/stores/responsiveMode'
 import type { HeadingLevel } from '~/utilities/html'
 
 type Props = {
-  rfc: Rfc
+  rfc: RfcCommon
   headingLevel?: HeadingLevel
 }
