+++ client/components/RFCCardDense.vue
@@ -21,7 +21,7 @@
     "
   >
     <template #headingTitle>
-      <component :is="formatTitle(`rfc${props.rfc.number}`)" />
+      <component :is="formatTitleAsVNode(`rfc${props.rfc.number}`)" />
     </template>
     <template #default>
       <RFCCardBody
@@ -34,14 +34,14 @@
 </template>
 
 <script setup lang="ts">
-import type { Rfc } from '../generated/red-client'
 import { infoRfcPathBuilder } from '../utilities/url'
-import { formatTitle } from '~/utilities/rfc'
+import { formatTitleAsVNode } from '~/utilities/rfc'
+import type { RfcCommon } from '~/utilities/rfc'
 import { useResponsiveModeStore } from '~/stores/responsiveMode'
 import type { HeadingLevel } from '~/utilities/html'
 
 type Props = {
-  rfc: Rfc
+  rfc: RfcCommon
   showAbstract?: boolean
   showTagDate?: boolean
   headingLevel?: HeadingLevel
