+++ client/utilities/rfc-index-txt.ts
@@ -2,8 +2,8 @@ import { DateTime } from 'luxon'
 import { padStart } from 'lodash-es'
 import { SPACE } from './strings'
 import { FIXME_getRFCMetadataWithMissingData } from './rfc.mocks'
-import { formatAuthor, formatIdentifiers } from './rfc'
 import { setTimeoutPromise } from './promises'
+import { formatAuthor, formatIdentifiers } from './rfc-converters-utils'
 import type { ApiClient, RfcMetadata } from '~/generated/red-client'
 
 // Note: this file is intentionally named rfc-index-txt.ts not rfc-index.txt.ts
